Move viewport-related functions to Viewport.cpp
