+++ src/openrct2/interface/Viewport.cpp
@@ -2023,6 +2023,233 @@ namespace OpenRCT2
             gameState.SavedViewRotation = viewport->rotation;
         }
     }
+
+    // converted from uint16_t values at 0x009A41EC - 0x009A4230
+    // these are percentage coordinates of the viewport to centre to, if a window is obscuring a location, the next is tried
+    // clang-format off
+    static constexpr float kWindowScrollLocations[][2] = {
+        { 0.5f, 0.5f },
+        { 0.75f, 0.5f },
+        { 0.25f, 0.5f },
+        { 0.5f, 0.75f },
+        { 0.5f, 0.25f },
+        { 0.75f, 0.75f },
+        { 0.75f, 0.25f },
+        { 0.25f, 0.75f },
+        { 0.25f, 0.25f },
+        { 0.125f, 0.5f },
+        { 0.875f, 0.5f },
+        { 0.5f, 0.125f },
+        { 0.5f, 0.875f },
+        { 0.875f, 0.125f },
+        { 0.875f, 0.875f },
+        { 0.125f, 0.875f },
+        { 0.125f, 0.125f },
+    };
+    // clang-format on
+
+    /**
+     *
+     *  rct2: 0x006E7C9C
+     * @param w (esi)
+     * @param x (eax)
+     * @param y (ecx)
+     * @param z (edx)
+     */
+    void WindowScrollToLocation(WindowBase& w, const CoordsXYZ& coords)
+    {
+        WindowUnfollowSprite(w);
+        if (w.viewport != nullptr)
+        {
+            int16_t height = TileElementHeight(coords);
+            if (coords.z < height - 16)
+            {
+                if (!(w.viewport->flags & VIEWPORT_FLAG_UNDERGROUND_INSIDE))
+                {
+                    w.viewport->flags |= VIEWPORT_FLAG_UNDERGROUND_INSIDE;
+                    w.Invalidate();
+                }
+            }
+            else
+            {
+                if (w.viewport->flags & VIEWPORT_FLAG_UNDERGROUND_INSIDE)
+                {
+                    w.viewport->flags &= ~VIEWPORT_FLAG_UNDERGROUND_INSIDE;
+                    w.Invalidate();
+                }
+            }
+
+            auto screenCoords = Translate3DTo2DWithZ(w.viewport->rotation, coords);
+
+            int32_t i = 0;
+            if (!(gScreenFlags & SCREEN_FLAGS_TITLE_DEMO))
+            {
+                bool found = false;
+                while (!found)
+                {
+                    auto x2 = w.viewport->pos.x + static_cast<int32_t>(w.viewport->width * kWindowScrollLocations[i][0]);
+                    auto y2 = w.viewport->pos.y + static_cast<int32_t>(w.viewport->height * kWindowScrollLocations[i][1]);
+
+                    auto it = WindowGetIterator(&w);
+                    for (; it != g_window_list.end(); it++)
+                    {
+                        auto w2 = (*it).get();
+                        auto x1 = w2->windowPos.x - 10;
+                        auto y1 = w2->windowPos.y - 10;
+                        if (x2 >= x1 && x2 <= w2->width + x1 + 20)
+                        {
+                            if (y2 >= y1 && y2 <= w2->height + y1 + 20)
+                            {
+                                // window is covering this area, try the next one
+                                i++;
+                                found = false;
+                                break;
+                            }
+                        }
+                    }
+                    if (it == g_window_list.end())
+                    {
+                        found = true;
+                    }
+                    if (i >= static_cast<int32_t>(std::size(kWindowScrollLocations)))
+                    {
+                        i = 0;
+                        found = true;
+                    }
+                }
+            }
+            // rct2: 0x006E7C76
+            if (w.viewport_target_sprite.IsNull())
+            {
+                if (!(w.flags & WF_NO_SCROLLING))
+                {
+                    w.savedViewPos = screenCoords
+                        - ScreenCoordsXY{ static_cast<int32_t>(w.viewport->ViewWidth() * kWindowScrollLocations[i][0]),
+                                          static_cast<int32_t>(w.viewport->ViewHeight() * kWindowScrollLocations[i][1]) };
+                    w.flags |= WF_SCROLLING_TO_LOCATION;
+                }
+            }
+        }
+    }
+
+    static void WindowViewportGetMapCoordsByCursor(
+        const WindowBase& w, int32_t* map_x, int32_t* map_y, int32_t* offset_x, int32_t* offset_y)
+    {
+        // Get mouse position to offset against.
+        auto mouseCoords = ContextGetCursorPositionScaled();
+
+        // Compute map coordinate by mouse position.
+        auto viewportPos = w.viewport->ScreenToViewportCoord(mouseCoords);
+        auto coordsXYZ = ViewportAdjustForMapHeight(viewportPos, w.viewport->rotation);
+        auto mapCoords = ViewportPosToMapPos(viewportPos, coordsXYZ.z, w.viewport->rotation);
+        *map_x = mapCoords.x;
+        *map_y = mapCoords.y;
+
+        // Get viewport coordinates centring around the tile.
+        int32_t z = TileElementHeight(mapCoords);
+
+        auto centreLoc = centre_2d_coordinates({ mapCoords.x, mapCoords.y, z }, w.viewport);
+        if (!centreLoc)
+        {
+            LOG_ERROR("Invalid location.");
+            return;
+        }
+
+        // Rebase mouse position onto centre of window, and compensate for zoom level.
+        int32_t rebased_x = w.viewport->zoom.ApplyTo(w.width / 2 - mouseCoords.x);
+        int32_t rebased_y = w.viewport->zoom.ApplyTo(w.height / 2 - mouseCoords.y);
+
+        // Compute cursor offset relative to tile.
+        *offset_x = w.viewport->zoom.ApplyTo(w.savedViewPos.x - (centreLoc->x + rebased_x));
+        *offset_y = w.viewport->zoom.ApplyTo(w.savedViewPos.y - (centreLoc->y + rebased_y));
+    }
+
+    static void WindowViewportCentreTileAroundCursor(
+        WindowBase& w, int32_t map_x, int32_t map_y, int32_t offset_x, int32_t offset_y)
+    {
+        // Get viewport coordinates centring around the tile.
+        int32_t z = TileElementHeight({ map_x, map_y });
+        auto centreLoc = centre_2d_coordinates({ map_x, map_y, z }, w.viewport);
+
+        if (!centreLoc.has_value())
+        {
+            LOG_ERROR("Invalid location.");
+            return;
+        }
+
+        // Get mouse position to offset against.
+        auto mouseCoords = ContextGetCursorPositionScaled();
+
+        // Rebase mouse position onto centre of window, and compensate for zoom level.
+        int32_t rebased_x = w.viewport->zoom.ApplyTo((w.width >> 1) - mouseCoords.x);
+        int32_t rebased_y = w.viewport->zoom.ApplyTo((w.height >> 1) - mouseCoords.y);
+
+        // Apply offset to the viewport.
+        w.savedViewPos = { centreLoc->x + rebased_x + w.viewport->zoom.ApplyInversedTo(offset_x),
+                           centreLoc->y + rebased_y + w.viewport->zoom.ApplyInversedTo(offset_y) };
+    }
+
+    void WindowZoomSet(WindowBase& w, ZoomLevel zoomLevel, bool atCursor)
+    {
+        Viewport* v = w.viewport;
+        if (v == nullptr)
+            return;
+
+        zoomLevel = std::clamp(zoomLevel, ZoomLevel::min(), ZoomLevel::max());
+        if (v->zoom == zoomLevel)
+            return;
+
+        // Zooming to cursor? Remember where we're pointing at the moment.
+        int32_t saved_map_x = 0;
+        int32_t saved_map_y = 0;
+        int32_t offset_x = 0;
+        int32_t offset_y = 0;
+        if (Config::Get().general.ZoomToCursor && atCursor)
+        {
+            WindowViewportGetMapCoordsByCursor(w, &saved_map_x, &saved_map_y, &offset_x, &offset_y);
+        }
+
+        // Zoom in
+        while (v->zoom > zoomLevel)
+        {
+            v->zoom--;
+            w.savedViewPos.x += v->ViewWidth() / 2;
+            w.savedViewPos.y += v->ViewHeight() / 2;
+        }
+
+        // Zoom out
+        while (v->zoom < zoomLevel)
+        {
+            v->zoom++;
+            w.savedViewPos.x -= v->ViewWidth() / 4;
+            w.savedViewPos.y -= v->ViewHeight() / 4;
+        }
+
+        // Zooming to cursor? Centre around the tile we were hovering over just now.
+        if (Config::Get().general.ZoomToCursor && atCursor)
+        {
+            WindowViewportCentreTileAroundCursor(w, saved_map_x, saved_map_y, offset_x, offset_y);
+        }
+
+        // HACK: Prevents the redraw from failing when there is
+        // a window on top of the viewport.
+        auto* windowMgr = Ui::GetWindowManager();
+        windowMgr->BringToFront(w);
+        w.Invalidate();
+    }
+
+    /**
+     * For all windows with viewports, ensure they do not have a zoom level less than the minimum.
+     */
+    void WindowCheckAllValidZoom()
+    {
+        WindowVisitEach([](WindowBase* w) {
+            if (w->viewport != nullptr && w->viewport->zoom < ZoomLevel::min())
+            {
+                WindowZoomSet(*w, ZoomLevel::min(), false);
+            }
+        });
+    }
 } // namespace OpenRCT2
 
 ZoomLevel ZoomLevel::min()
